backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.21.0
nltk>=3.7
joblib>=1.2.0
python-dotenv>=0.19.0